requirements.txt
flask==3.1.0
torch==2.7.0
torchvision==0.22.0
numpy==2.2.5
pillow==11.2.1
opencv-python==4.11.0.86
pytesseract==0.3.13
gunicorn==20.1.0 
fastapi
uvicorn
torch
pillow
python-multipart
